+++ packages/uilib/src/lib/plugins/communication-explorer/sidebar/ied-accordion/index.ts
@@ -1,1 +1,3 @@
 export {default as IEDAccordion} from "./ied-accordion.svelte"
+export * from "./types"
+export * from "./groupRelationsByServiceType"