+++ packages/uilib/src/lib/plugins/communication-explorer/sidebar/ied-accordion/types.ts
@@ -1,0 +1,10 @@
+import type { MessageType } from "@oscd-plugins/core"
+import type { IEDNode } from "../../../../components/diagram/nodes"
+
+export type ServiceTypeGroup = Map<string, ServiceObject[]>;
+
+export type ServiceObject = {
+    node: IEDNode;
+    serviceType: MessageType | "Unknown";
+    serviceTypeLabel?: string;
+};