+++ packages/uilib/src/lib/plugins/documentation/documentation-print-view/print-documentation/print-bays/print-ieds/print-ieds.svelte
@@ -1,12 +1,48 @@
 <script lang="ts">
-    import type { IEDCommInfo } from "@oscd-plugins/core"
+    import type { IEDCommInfo, ReceivedMessage } from "@oscd-plugins/core"
     import { PrintMessageServiceChip } from "../../../../../../components/print-message-service-chip"
-    import { convertIEDNameToID } from "../../_shared-functions"
+    import { convertIEDNameToID as convertIEDNameToHrefID } from "../../_shared-functions"
+    import type { IEDConnection } from "../_shared-functions"
+    import { Icons, openSCDIcons } from "../../../../../../components/icons"
 
-    export let publishedServiceTypes: string[] | undefined = undefined
+    export let publishedServiceTypes: IEDConnection[] | undefined = undefined
     export let ied: IEDCommInfo
 
-    const gotoIedNameID = convertIEDNameToID(ied.iedName, false)
+    $: groupedReceiverByServiceTypeID = groupByServiceTypeIDs(ied.received)
+
+    type GroupedServiceType = {
+        serviceType: string,
+        serviceTypeLabel: string,
+        received: ReceivedMessage,
+    }
+
+    export function groupByServiceTypeIDs( relations: ReceivedMessage[] ) {
+
+    	const array = new Map<string, GroupedServiceType[]>()
+
+    	relations.forEach((element) => {
+    		let serviceType: string = element.serviceType
+    		let serviceTypeID: string = element.srcCBName
+
+    		const keyName = `${serviceType}_${serviceTypeID}`
+    		const content: GroupedServiceType = {
+    			serviceType:      serviceType,
+    			serviceTypeLabel: serviceTypeID,
+    			received:         element,
+    		}
+
+    		const hasServiceTypeElement = array.has(keyName)
+    		if (!hasServiceTypeElement) {
+    			array.set(keyName, []) 
+    		}
+
+    		const serviceTypeElement = array.get(keyName)
+    		serviceTypeElement?.push(content)
+    	})
+    	return array
+    }
+
+    const gotoIedNameID = convertIEDNameToHrefID(ied.iedName, false)
 </script>
 
 <div class="ied-node" id={gotoIedNameID}>
@@ -20,7 +56,7 @@
                 {#each publishedServiceTypes as serviceType}
                     <li>
                         <PrintMessageServiceChip
-                            type={serviceType.toLocaleLowerCase()}
+                            type={ serviceType.serviceType }
                         />
                     </li>
                 {/each}
@@ -30,37 +66,39 @@
     {#if ied.received.length > 0}
         <div>
             <h5>Subscriber</h5>
-            <ul class="subscriber-list">
-                {#each ied.received as subscriber}
-                    <li>
-                        <div class="chip">
-                            <PrintMessageServiceChip
-                                type={subscriber.serviceType.toLocaleLowerCase()}
-                            />
+            <div class="services">
+                {#each Array.from(groupedReceiverByServiceTypeID) as serviceTypeGroup}
+                    {@const receiver = serviceTypeGroup[1]}
+                    {@const serviceTyp = receiver[0]?.serviceType ?? "Unknown"}
+                    {@const serviceTypLabel = receiver[0]?.serviceTypeLabel ?? "Unknown"}
+
+                    <div class="service">
+                        <div class="service-type-grouper service-{serviceTyp.toLocaleLowerCase()}">
+                            <p>{`${serviceTyp} - ${serviceTypLabel}`}</p>
+                            <p>ID: {serviceTypLabel}</p>
+                            <p>ServiceType: {serviceTyp}</p>
                         </div>
-                        <a href={convertIEDNameToID(subscriber.iedName, true)}>
-                            <span>
-                                {subscriber.iedName}
-                            </span>
-                        </a>
-                    </li>
-                {/each}
-            </ul>
+                        <ul class="subscriber-list">
+                            {#each receiver as record} 
+                                {@const connectedIed = record.received.iedName}
+                                <li>
+                                    <div class="iedConnector service-{serviceTyp.toLocaleLowerCase()}">
+                                        <span>{connectedIed}</span>
+                                        <a href={convertIEDNameToHrefID(connectedIed, true)}>
+                                            <Icons name={"linkIcon"} size="normal" />
+                                        </a>
+                                    </div>
+                                </li>
+                            {/each}
+                        </ul>
+                    </div>
+                    {/each}
+            </div>
         </div>
     {/if}
 </div>
 
 <style lang="scss">
-    ul.subscriber-list {
-        display: flex;
-        flex-direction: column;
-        gap: 0.2rem;
-        li {
-            display: flex;
-            flex-direction: row;
-            gap: 0.3rem;
-        }
-    }
 
     .show-publiished-service-types {
         display: flex;
@@ -69,4 +107,60 @@
         gap: 0.5rem;
         list-style-type: none;
     }
+
+    .services {
+        display: flex; 
+        flex-direction: column;
+        gap: 1rem;     
+
+        .service {
+            display: flex; 
+            gap: 1rem;     
+    
+            .service-type-grouper {
+                border: 2px solid #000;
+                border-radius: 5px;
+                padding: .5rem;
+                width: 30%;
+        
+                p {
+                    margin: 0;
+                    padding: 0;
+                }
+
+                &.service-goose {
+                    border-color: var(--color-message-goose);
+                }
+                &.service-mms, .service-reportcontrol {
+                    border-color: var(--color-message-mms);
+                }
+                &.service-sampledvalues, .service-smv {
+                    border-color: var(--color-message-sampledvalues);
+                }
+            }
+
+            ul.subscriber-list {
+                display: flex;
+                flex-direction: column;
+                gap: 0.2rem;
+                padding: 0;
+                li {
+                    display: flex;
+                    flex-direction: row;
+                    gap: 0.3rem;
+                }
+            }
+
+            .iedConnector {
+                display: flex; 
+                align-items: center; 
+                gap: .5rem;
+                border: 1px solid black;
+                padding: .2rem 1rem;
+                border-radius: 5px;
+            }
+        }
+    }
+
+
 </style>